исправил замечания, разобрался с тестами.
